# openapi.yaml
openapi: 3.0.3
info:
  title: EduFlow API
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/plans:
    get:
      summary: List plans
      operationId: listPlans
      security: [{ betterAuth: [] }]
      responses:
        '200':
          {
            description: Plans,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        required: [data],
                        properties:
                          { data: { type: array, items: { $ref: '#/components/schemas/Plan' } } },
                      },
                  },
              },
          }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
    post:
      summary: Create a plan
      operationId: createPlan
      security: [{ betterAuth: [] }]
      requestBody:
        {
          required: true,
          content: { application/json: { schema: { $ref: '#/components/schemas/PlanInput' } } },
        }
      responses:
        '201': { description: Plan created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
        '409': { description: Plan slug already exists }
...
  /api/v1/plans/{id}:
    patch:
      summary: Update a plan
      operationId: updatePlan
      security: [{ betterAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      requestBody:
        {
          required: true,
          content: { application/json: { schema: { $ref: '#/components/schemas/PlanInput' } } },
        }
      responses:
        '200': { description: Plan updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
        '404': { description: Plan not found }
        '409': { description: Plan slug already exists }
    delete:
      summary: Delete a plan
      operationId: deletePlan
      security: [{ betterAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      responses:
        '200':
          description: Plan deleted
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [id, name, slug]
                    properties:
                      id: { type: string, format: uuid }
                      name: { type: string }
                      slug: { type: string }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
        '404': { description: Plan not found }
        '409': { description: Plan in use by subscriptions or payments }
  /api/v1/plans/features:
    get:
      summary: List the Platform feature catalog for plan configuration
      operationId: listPlatformFeatures
      security: [{ betterAuth: [] }]
      responses:
        '200':
          description: Feature catalog
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/FeatureCatalogItem' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
...
  /api/v1/payment-requests:
    post:
      summary: Submit a manual subscription payment
      operationId: createPaymentRequest
      security:
        - betterAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Pending payment request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400': { description: Validation error, unavailable plan, or amount mismatch }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/WorkspaceMembershipRequired' }
        '409': { $ref: '#/components/responses/DuplicateTransaction' }
  /api/v1/payment-requests/account:
    post:
      summary: Submit a manual subscription payment for the authenticated account
      operationId: createAccountPaymentRequest
      description: Creates an account-level pending purchase without creating a workspace. Platform Owner approval unlocks one workspace creation.
      security: [{ betterAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePaymentRequest' }
      responses:
        '201':
          description: Pending account payment request created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentResponse' }
        '400': { description: Validation error, unavailable plan, or amount mismatch }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { $ref: '#/components/responses/DuplicateTransaction' }
...
  /api/v1/payment-requests/pending:
    get:
      summary: List pending payment requests
      operationId: listPendingPaymentRequests
      security:
        - betterAuth: []
      parameters:
        - name: search
          in: query
          description: Case-insensitive search across workspace, plan, sender number, transaction ID, and requester ID.
          schema: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '200':
          description: Pending payment requests
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PendingPaymentRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
  /api/v1/payment-requests/revenue-overview:
    get:
      summary: Get Platform Owner payment and revenue aggregates
      operationId: getRevenueOverview
      description: Aggregates payment records in PostgreSQL and returns exact minor-unit amounts as decimal strings.
      security: [{ betterAuth: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          {
            description: Exact payment aggregates,
            approved revenue by plan,
            and recent payment activity,
          }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/PlatformOwnerRequired' }
...
components:
  securitySchemes:
    betterAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
...
    PlanId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    WorkspaceId:
      name: X-Workspace-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    PlanInput:
      type: object
      additionalProperties: false
      required: [name, slug, priceMinor, durationDays, trialDays]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        priceMinor: { oneOf: [{ type: integer, minimum: 0 }, { type: string, pattern: '^\\d+$' }] }
        durationDays: { type: integer, minimum: 0 }
        trialDays: { type: integer, minimum: 0 }
        isActive: { type: boolean }
        features:
          {
            type: array,
            items:
              {
                type: object,
                required: [featureKey, enabled, limitValue],
                properties:
                  {
                    featureKey: { type: string },
                    enabled: { type: boolean },
                    limitValue:
                      {
                        oneOf: [{ type: integer, minimum: 0 }, { type: string, pattern: '^\\d+$' }],
                        nullable: true,
                      },
                  },
                additionalProperties: false,
              },
          }
    Plan:
      allOf:
        - $ref: '#/components/schemas/PlanInput'
        - type: object
          properties:
            {
              id: { type: string, format: uuid },
              priceMinor: { type: string },
              createdAt: { type: string, format: date-time },
              updatedAt: { type: string, format: date-time },
            }
    FeatureCatalogItem:
      type: object
      required: [key, name, description]
      properties:
        key: { type: string, maxLength: 100 }
        name: { type: string, maxLength: 120 }
        description: { type: string, nullable: true }
...
    CreatePaymentRequest:
      type: object
      additionalProperties: false
      required: [planId, amount, paymentMethod, senderNumber, transactionId]
      properties:
        planId: { type: string, format: uuid }
        amount: { oneOf: [{ type: integer, minimum: 0 }, { type: string, pattern: '^\\d+$' }] }
        paymentMethod: { type: string, enum: [cash, bkash, nagad, rocket, other] }
        senderNumber: { type: string, minLength: 1, maxLength: 30 }
        transactionId: { type: string, minLength: 1, maxLength: 100 }
    PaymentRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        planId: { type: string, format: uuid, nullable: true }
        amountMinor: { type: string, pattern: '^\\d+$' }
        paymentMethod: { type: string, enum: [cash, bkash, nagad, rocket, other] }
        senderNumber: { type: string }
        transactionId: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        reviewedAt: { type: string, format: date-time, nullable: true }
        reviewedByUserId: { type: string, nullable: true }
        rejectionReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    PendingPaymentRequest:
      type: object
      required:
        [
          id,
          purpose,
          amountMinor,
          paymentMethod,
          senderNumber,
          transactionId,
          status,
          createdAt,
          workspace,
          plan,
        ]
      properties:
        id: { type: string, format: uuid }
        purpose: { type: string, enum: [subscription, sms_credit] }
        amountMinor: { type: string, pattern: '^\\d+$' }
        paymentMethod: { type: string, enum: [cash, bkash, nagad, rocket, other] }
        senderNumber: { type: string }
        transactionId: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        createdAt: { type: string, format: date-time }
        workspace:
          type: object
          required: [id, name, slug]
          properties:
            id: { type: string, format: uuid }
            name: { type: string, nullable: true }
            slug: { type: string, nullable: true }
        plan:
          type: object
          nullable: true
          required: [id, name, slug]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            slug: { type: string }
    PaymentResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/PaymentRequest' }
...
  responses:
...
    ValidationError:
      description: Invalid request
    Unauthenticated:
      description: Authentication required
    WorkspaceMembershipRequired:
      description: Workspace membership required
    PlatformOwnerRequired:
      description: Platform Owner authorization required
    DuplicateTransaction:
      description: Transaction ID already exists
...